requirements.txt
Flask
python-docx
PyPDF2
chardet
werkzeug
toml
